Created a reusable confirmation component which fires when the user clicks Delete button on a task. After that a modal shows up, and the user can confirm if he or she wants to delete that Task.

+++ FlowTaskManager.Web/Client/Pages/TaskPages/DisplayTaskBase.cs
@@ -1,8 +1,10 @@
-﻿using FlowTaskManager.Web.Client.Services;
+﻿using FlowTaskManager.Web.Client.Modals;
+using FlowTaskManager.Web.Client.Services;
 using FlowTaskManager.Web.Shared.Models;
 using Microsoft.AspNetCore.Components;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -16,11 +18,26 @@ namespace FlowTaskManager.Web.Client.Pages.TaskPages
         private NavigationManager NavigationManager { get; set; }
         [Parameter]
         public ProgrammingTask Task { get; set; }
-
+        public ConfirmationBase DeleteConfirmation { get; set; }
+        [Parameter]
+        public EventCallback<bool> TaskDeletion { get; set; }
+        
         public async void ClickDelete()
         {
-            await TaskService.DeleteProgrammingTask(Task.Id);
-            NavigationManager.NavigateTo("/tasks", true);
+            DeleteConfirmation.Title = "Delete Task";
+            DeleteConfirmation.Content = "Are you sure you want to delete this task?";
+            DeleteConfirmation.ActionButton = "Delete";
+            DeleteConfirmation.CancelButton = "Cancel";
+            DeleteConfirmation.Show();
+        }
+
+        public async void DeleteModalResult(bool actionValue)
+        {
+            if (actionValue)
+            {
+                await TaskService.DeleteProgrammingTask(Task.Id);
+                await TaskDeletion.InvokeAsync(true);
+            }
         }
     }
 }
